--- neuwon/api/__init__.py
@@ -1,15 +1,0 @@
-__all__ = """
-Location
-Model
-Neighbor
-Reaction
-Real
-Species
-""".split()
-
-from neuwon.common import Real, Location
-from neuwon.model import Model
-from neuwon.reactions import Reaction
-from neuwon.segments import Segment
-from neuwon.species import Species
-from neuwon.voronoi import Neighbor
