--- neuwon/species.py
@@ -1,181 +1,0 @@
-import numpy as np
-from scipy.sparse import csr_matrix, csc_matrix
-from scipy.sparse.linalg import expm
-import cupy as cp
-import cupyx.scipy.sparse
-import math
-import copy
-from collections.abc import Callable, Iterable, Mapping
-from neuwon.common import Real, epsilon
-
-F = 96485.3321233100184 # Faraday's constant, Coulombs per Mole of electrons
-R = 8.31446261815324 # Universal gas constant
-
-library = {
-    "na": {
-        "charge": 1,
-        "transmembrane": True,
-        "reversal_potential": "nerst",
-        "intra_concentration":  15e-3,
-        "extra_concentration": 145e-3,
-    },
-    "k": {
-        "charge": 1,
-        "transmembrane": True,
-        "reversal_potential": "nerst",
-        "intra_concentration": 150e-3,
-        "extra_concentration":   4e-3,
-    },
-    "ca": {
-        "charge": 2,
-        "transmembrane": True,
-        "reversal_potential": "goldman_hodgkin_katz",
-        "intra_concentration": 70e-9,
-        "extra_concentration": 2e-3,
-    },
-    "cl": {
-        "charge": -1,
-        "transmembrane": True,
-        "reversal_potential": "nerst",
-        "intra_concentration":  10e-3,
-        "extra_concentration": 110e-3,
-    },
-    "glu": {
-        # "extra_concentration": 1/0, # TODO!
-        "extra_diffusivity": 1e-6, # TODO!
-        # "extra_decay_period": 1/0, # TODO!
-    },
-}
-
-class _Diffusion:
-    def __init__(self, time_step, geometry, species, where):
-        self.time_step = time_step
-        # Compute the coefficients of the derivative function:
-        # dX/dt = C * X, where C is Coefficients matrix and X is state vector.
-        cols = [] # Source
-        rows = [] # Destintation
-        data = [] # Weight
-        # derivative(Destintation) += Source * Weight
-        if where == "intracellular":
-            for location in range(len(geometry)):
-                if geometry.is_root(location):
-                    continue
-                parent = geometry.parents[location]
-                l = geometry.lengths[location]
-                flux = species.intra_diffusivity * geometry.cross_sectional_areas[location] / l
-                cols.append(location)
-                rows.append(parent)
-                data.append(+1 * flux / geometry.intra_volumes[parent])
-                cols.append(location)
-                rows.append(location)
-                data.append(-1 * flux / geometry.intra_volumes[location])
-                cols.append(parent)
-                rows.append(location)
-                data.append(+1 * flux / geometry.intra_volumes[location])
-                cols.append(parent)
-                rows.append(parent)
-                data.append(-1 * flux / geometry.intra_volumes[parent])
-            for location in range(len(geometry)):
-                cols.append(location)
-                rows.append(location)
-                data.append(-1 / species.intra_decay_period)
-        elif where == "extracellular":
-            D = species.extra_diffusivity / geometry.extracellular_tortuosity ** 2
-            for location in range(len(geometry)):
-                for neighbor in geometry.neighbors[location]:
-                    flux = D * neighbor["border_surface_area"] / neighbor["distance"]
-                    cols.append(location)
-                    rows.append(neighbor["location"])
-                    data.append(+1 * flux / geometry.extra_volumes[neighbor["location"]])
-                    cols.append(location)
-                    rows.append(location)
-                    data.append(-1 * flux / geometry.extra_volumes[location])
-            for location in range(len(geometry)):
-                cols.append(location)
-                rows.append(location)
-                data.append(-1 / species.extra_decay_period)
-        # Note: always use double precision floating point for building the impulse response matrix.
-        coefficients = csc_matrix((data, (rows, cols)), shape=(len(geometry), len(geometry)), dtype=float)
-        coefficients.data *= self.time_step
-        self.irm = expm(coefficients)
-        # Prune the impulse response matrix at epsilon nanomolar (mol/L).
-        self.irm.data[np.abs(self.irm.data) < epsilon * 1e-6] = 0
-        self.irm.eliminate_zeros()
-        if True: print(where, species.name, "IRM NNZ per Location", self.irm.nnz / len(geometry))
-        self.irm = cupyx.scipy.sparse.csr_matrix(self.irm, dtype=Real)
-
-def nerst_potential(charge, T, intra_concentration, extra_concentration):
-    """ Returns the reversal voltage for an ionic species. """
-    xp = cp.get_array_module(intra_concentration)
-    if charge == 0: return xp.full_like(intra_concentration, xp.nan)
-    ratio = xp.divide(extra_concentration, intra_concentration)
-    return xp.nan_to_num(R * T / F / charge * np.log(ratio))
-
-@cp.fuse()
-def _efun(z):
-    if abs(z) < 1e-4:
-        return 1 - z / 2
-    else:
-        return z / (math.exp(z) - 1)
-
-def goldman_hodgkin_katz(charge, T, intra_concentration, extra_concentration, voltages):
-    """ Returns the reversal voltage for an ionic species. """
-    xp = cp.get_array_module(intra_concentration)
-    if charge == 0: return xp.full_like(intra_concentration, np.nan)
-    z = (charge * F / (R * T)) * voltages
-    return (charge * F) * (intra_concentration * _efun(-z) - extra_concentration * _efun(z))
-
-class _Electrics:
-    def __init__(self, time_step, geometry,
-            intracellular_resistance = 1,
-            membrane_capacitance = 1e-2,
-            initial_voltage = -70e-3):
-        # Save and check the arguments.
-        self.time_step                  = time_step
-        self.intracellular_resistance   = float(intracellular_resistance)
-        self.membrane_capacitance       = float(membrane_capacitance)
-        assert(self.intracellular_resistance > 0)
-        assert(self.membrane_capacitance > 0)
-        # Initialize data buffers.
-        self.voltages           = cp.full(len(geometry), initial_voltage, dtype=Real)
-        # Compute passive properties.
-        self.axial_resistances  = np.empty(len(geometry), dtype=Real)
-        self.capacitances       = np.empty(len(geometry), dtype=Real)
-        for location in range(len(geometry)):
-            l = geometry.lengths[location]
-            sa = geometry.surface_areas[location]
-            xa = geometry.cross_sectional_areas[location]
-            self.axial_resistances[location] = self.intracellular_resistance * l / xa
-            self.capacitances[location] = self.membrane_capacitance * sa
-        # Compute the coefficients of the derivative function:
-        # dX/dt = C * X, where C is Coefficients matrix and X is state vector.
-        cols = [] # Source
-        rows = [] # Destintation
-        data = [] # Weight
-        for location in range(len(geometry)):
-            if geometry.is_root(location):
-                continue
-            parent = geometry.parents[location]
-            r = self.axial_resistances[location]
-            cols.append(location)
-            rows.append(parent)
-            data.append(+1 / r / self.capacitances[parent])
-            cols.append(location)
-            rows.append(location)
-            data.append(-1 / r / self.capacitances[location])
-            cols.append(parent)
-            rows.append(location)
-            data.append(+1 / r / self.capacitances[location])
-            cols.append(parent)
-            rows.append(parent)
-            data.append(-1 / r / self.capacitances[parent])
-        # Note: always use double precision floating point for building the impulse response matrix.
-        coefficients = csc_matrix((data, (rows, cols)), shape=(len(geometry), len(geometry)), dtype=np.float64)
-        coefficients.data *= self.time_step
-        self.irm = expm(coefficients)
-        # Prune the impulse response matrix at epsilon millivolts.
-        self.irm.data[np.abs(self.irm.data) < epsilon * 1e-3] = 0
-        self.irm.eliminate_zeros()
-        if True: print("Electrics IRM NNZ per Location", self.irm.nnz / len(geometry))
-        # Move this data to the GPU now that the CPU is done with it.
-        self.irm = cupyx.scipy.sparse.csr_matrix(self.irm, dtype=Real)
